+++ ChangeLog.md
@@ -17,4 +17,11 @@
 ### Modification
 
 - Correction de l'affichage des puces de la map
- 
+
+## [2.0.0] - 2025-09-07
+
+### Modification
+
+- Ajout de la connexion
+- Ajout de la création de compte
+- Ajout de la modification des données utilisateurs